Refactor: Move filter types to a separate directory
This commit moves the `StringFilter`, `RangeFilter`, and `Filter` classes to the `types` directory.
Also updates all references to the moved classes.
