+++ lib/widgets/overview_page/filters/active_filters.dart
@@ -1,6 +1,6 @@
-import 'package:exdock_backend_client/widgets/overview_page/filters/filter.dart';
 import 'package:exdock_backend_client/widgets/overview_page/filters/filter_notifier.dart';
 import 'package:exdock_backend_client/widgets/overview_page/filters/filter_widget.dart';
+import 'package:exdock_backend_client/widgets/overview_page/filters/types/filter.dart';
 import 'package:flutter/material.dart';
 
 class ActiveFilters extends StatefulWidget {
