+++ lib/widgets/overview_page/filters/filter_notifier.dart
@@ -1,4 +1,4 @@
-import 'package:exdock_backend_client/widgets/overview_page/filters/filter.dart';
+import 'package:exdock_backend_client/widgets/overview_page/filters/types/filter.dart';
 import 'package:flutter/material.dart';
 
 class FilterNotifier extends ValueNotifier<Map<String, Filter>> {
