+++ lib/widgets/overview_page/filters/types/range_filter.dart
@@ -1,4 +1,4 @@
-import 'package:exdock_backend_client/widgets/overview_page/filters/filter.dart';
+import 'package:exdock_backend_client/widgets/overview_page/filters/types/filter.dart';
 
 class RangeFilter extends Filter {
   RangeFilter({
